requirements.txt
streamlit==1.45.0
supabase==2.15.1
pandas==2.2.3
google-auth-oauthlib==1.2.2  
gspread==5.7.2
google-auth==2.24.0
oauth2client==4.1.3
pytesseract==0.3.10
Pillow==9.5.0
